// frontend/src/app/api/generated/model/categoryUpdate.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface CategoryUpdate {
  description?: string | null;
  is_active?: boolean | null;
  name?: string | null;
}