// frontend/src/app/api/generated/model/passwordResetRequestResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PasswordResetUserResponse } from './passwordResetUserResponse';
import { PasswordResetRequestStatus } from './passwordResetRequestStatus';

export interface PasswordResetRequestResponse {
  id: number;
  requested_at: string;
  resolved_at: string | null;
  resolved_by_user_id: number | null;
  status: PasswordResetRequestStatus;
  user: PasswordResetUserResponse;
}