// frontend/src/app/api/generated/model/userProvisionRequest.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRole } from './userRole';

export interface UserProvisionRequest {
  email: string;
  first_name: string;
  last_name: string;
  password: string;
  role: UserRole;
}