// frontend/src/app/api/generated/model/userResponse.ts
/**
 * PulseDesk API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { UserRole } from './userRole';

export interface UserResponse {
  created_at: string;
  email: string;
  first_name: string;
  id: number;
  is_active: boolean;
  last_name: string;
  role: UserRole;
  updated_at: string;
}